+++ src/sections/HeroCarousel/HeroCarousel.module.scss
@@ -1,4 +1,4 @@
-@use '../stylesheets/modules/colors';
+@use '../../stylesheets/modules/colors';
 
 .flex {
   display: flex;
