+++ src/sections/HeroCarousel/HeroCarousel.tsx
@@ -1,14 +1,13 @@
-import React, { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react';
+import React from 'react';
 import styles from './HeroCarousel.module.scss';
 import { Carousel } from 'react-responsive-carousel';
 import { uniqueId } from 'lodash';
-import Img from 'react-cool-img';
 import 'react-responsive-carousel/lib/styles/carousel.min.css'; // requires a loader
-import { placeholderImg } from '../assets';
-import { Button } from '../components/Buttons';
-import IconButton, { IIconButtonProps } from '../components/Buttons/IconButton';
+import { placeholderImg } from '../../assets';
+import { Button } from '../../components/Buttons';
+import IconButton, { IIconButtonProps } from '../../components/Buttons/IconButton';
 import { useQuery } from '@apollo/client';
-import GET_PAGE_DATA, { IPageData } from '../data/get-data';
+import GET_PAGE_DATA, { IPageData } from '../../data/get-data';
 import { useResizeObserver } from '@mantine/hooks';
 
 const socialMedia: IIconButtonProps[] = [
